// enums/wow-enums.ts
export enum WowRegion {
    EU = "EU",
    US = "US",
    KR = "KR",
    CN = "CN",
    TW = "TW",
}

export enum WowFaction {
    ALLIANCE = "Alliance",
    HORDE = "Horde",
}

export enum WowVersion {
    SOD = "SoD",
}

export enum WowSodRealm {
    CRUSADER_STRIKE = "Crusader Strike",
    LONE_WOLF = "Lone Wolf",
    LIVING_FLAME = "Living Flame",
    WILD_GROWTH = "Wild Growth",
    SHADOWSTRIKE = "Shadowstrike",
    CHAOS_BOLT = "Chaos Bolt",
    LAVA_LASH = "Lava Lash",
    PENANCE = "Penance",
}